requirements.txt
absl-py==2.3.0
altair==5.5.0
anyio==4.9.0
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
asttokens==3.0.0
astunparse==1.6.3
async-lru==2.0.5
attrs==25.3.0
babel==2.17.0
beautifulsoup4==4.13.4
bleach==6.2.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
comm==0.2.2
contourpy==1.3.0
cycler==0.12.1
debugpy==1.8.14
decorator==5.2.1
defusedxml==0.7.1
exceptiongroup==1.3.0
executing==2.2.0
fastjsonschema==2.21.1
flatbuffers==25.2.10
fonttools==4.58.4
gast==0.4.0
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.40.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.73.1
h11==0.16.0
h5py==3.14.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
importlib-metadata==8.7.0
importlib-resources==6.5.2
ipykernel==6.29.5
ipython==8.18.1
ipywidgets==8.1.7
jax==0.4.30
jaxlib==0.4.30
jedi==0.19.2
jinja2==3.1.6
joblib==1.5.1
json5==0.12.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter==1.1.1
jupyter-client==8.6.3
jupyter-console==6.6.3
jupyter-core==5.8.1
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter-server==2.16.0
jupyter-server-terminals==0.5.3
jupyterlab==4.4.4
jupyterlab-pygments==0.3.0
jupyterlab-server==2.27.3
jupyterlab-widgets==3.0.15
keras==2.12.0
kiwisolver==1.4.7
libclang==18.1.1
markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.4
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.3
ml-dtypes==0.5.1
namex==0.1.0
narwhals==1.44.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
notebook==7.4.4
notebook-shim==0.2.4
numpy==1.23.5
oauthlib==3.3.1
opencv-python==4.11.0.86
opt-einsum==3.4.0
optree==0.16.0
overrides==7.7.0
packaging==25.0
pandas==2.3.0
pandocfilters==1.5.1
parso==0.8.4
pillow==11.2.1
platformdirs==4.3.8
prometheus-client==0.22.1
prompt-toolkit==3.0.51
protobuf==4.25.8
psutil==7.0.0
pure-eval==0.2.3
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycparser==2.22
pydeck==0.9.1
pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-json-logger==3.3.0
pytz==2025.2
pywin32==310
pywinpty==2.0.15
PyYAML==6.0.2
pyzmq==27.0.0
referencing==0.36.2
requests==2.32.4
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==14.0.0
rpds-py==0.25.1
rsa==4.9.1
scikit-learn==1.6.1
scipy==1.13.1
seaborn==0.13.2
Send2Trash==1.8.3
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
stack-data==0.6.3
streamlit==1.46.1
tenacity==9.1.2
tensorboard==2.12.3
tensorboard-data-server==0.7.2
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-intel==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.1.0
terminado==0.18.1
threadpoolctl==3.6.0
tinycss2==1.4.0
toml==0.10.2
tomli==2.2.1
tornado==6.5.1
traitlets==5.14.3
typing-extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
watchdog==6.0.0
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.8.0
werkzeug==3.1.3
widgetsnbextension==4.0.14
wrapt==1.14.1
zipp==3.23.0
